Run Kubeval tools separately in Docker container
This does not have a significant impact on the performance while at the
same time, it allows to move the code outside of the string.

This way, the syntax can be highlighted properly, static code analysis
tools can be run and testing is much easier. These points are crucial to
improve the code quality in terms of readability and error detection or
prevention.

As per best practices to avoid conflicts, the variables in the
introduced function are declared local and their names are lower case.