Wrap script in main function
This allows to declare functions below the places where they are first
called. This way, the script can be read in a natural way from top to
bottom with reduced visual jumping where one first reads what is done on
a more abstract level and then goes on to read more details below.