Use Bash instead of Bourne shell
The Bourne shell was the first shell and is the lowest common
denominator of modern shells. Bash is a popular shell that adds many
features improving readability and efficiency, therefore it is
preferred.

One reason to target the Bourne shell would be portability of the script
code between different systems, but this is not a requirement here.