Add initial test for successful script execution
For "true" unit tests, the single functions of a script should be
tested. For that, however, either the script would have to be split into
one containing the functions and one with the actual call or add a line
checking for a special ShellSpec variable. Both options seem a bit
invasive for such a small script, so sourcing it has been chosen as the
next best option.

The test verifies a successful exit status and the occurrence of at
least one pass message in the output. The output is also printed to
visually check that the test passes correctly.