Test that there are no generated files left over
Currently, a command executed in Docker creates a file that is processed
by following commands. Although the file is not required anymore after
that, it is not deleted. This pollutes the file hierarchy with a
temporary file that in addition belongs to the root user because of its
creation from inside the container which makes it tedious to remove.

Therefore, a test is added that reflects the desired state of no
additional files left after script execution. ShellSpec's file assertion
facilities are used for that.