Test for colour codes in output
To be able to refactor the way coloured text is output and make sure the
functionality isn't unintentionally changed, the colour codes are
included in the output assertions of related tests.

ShellSpec's functionality to match the output against a pattern with
wildcards is used to verify that the text surrounded the colour limiters
contains the word "error".