Add mock for Find command
Find interacts with the file system, so if we used it in the tests as
is, we would have to create the necessary files and remove them for each
test. We don't want to write such elaborated integration tests, but
focus on rather simple unit tests here, so we mock Find.