Fix processing cluster names
Normally, Find's `-exec` feature should be used to process the found
entries. This avoids issues with word splitting that occur when the
processing is done via a loop.

However, in this script we have the requirement of first printing all
found files, then processing each of them. One idea could be to provide
one execution parameter operating on all files at once (terminated with
`+`) for printing and a second operating on single files (terminated
with `;`) for the processing. This doesn't work because Find executes
the aggregating operation last.

So instead of calling Find twice for the two operations, each entry is
terminated with a character that isn't allowed in filenames. That
enables storing them correctly in an array for both the output and the
processing.

The array has to be declared separately via builtin to make it a local
variable.

Process substitution has to be used to pass the output of Find to
Readarray. As the process is executed in a background job, we wait on
the process ID retrieved via special parameter to get its exit status
and abort if an error occurred. We also have to group the conditional
operations to avoid unexpected interpretation.

The variable containing the found entries has been renamed to lower case
because upper case is typically used for the system's or shell's
variables.