Remove useless use of cat
Using Cat to pipe the content of one file to other commands goes against
the Unix philosophy in that this command is meant to do one thing which
is to concatenate file contents.

It is also often unnecessary as commands accept a filename as parameter
and if not, file redirection exists exactly for the purpose of piping
file contents.

Piping via Cat can also lead to loss of information and therefore
functionality down the pipe.

A popular counter-argument is that with Cat, the data flow can be read
from left to right. However, this can also be achieved with file
redirection as it can be put before the reading command.