Replace remaining Echo calls containing colour codes
The Echo calls to display an error and exit can be replaced with the
previously created function for that purpose.

This also removes the unnecessarily indirect checks of the exit status
via special variable and the use of `[ ]` tests which are inferior to
Bash's `[[ ]]` tests. For the last validation step, an arithmetic
conditional construct is used as replacement.

To interpret the colour escape codes with Printf, `%b` could be used
instead of `%s`, but the latter in combination with ANSI-C quotes is
preferred for consistency.

An optional exit status parameter is added to the exit function via a
parameter expansion that sets the last exit status as default.