Test for exact number of passed validations
A function is added to use ShellSpec's functionality of defining custom
assertions. The output as the test subject is passed in a parameter of
the same name as the function and a parameter expansion is used to exit
if it is missing.

The output variable is then passed to Grep via here string to determine
the number of occurrences of the pass message.

The conditional construct for arithmetic expressions is then used to
verify if the number is the expected one.