Generate temporary file and delete it after execution
Mktemp is used to create the temporary file. The advantage over a fixed
filename is that race conditions are avoided due to the randomly varying
filename for each execution.

Note that the error check can't be part of the function used via command
substitution because it would only exit the subshell, not the script.

A trap is used to remove the temporary file when the script stops.

The output redirection to the file has been moved from inside the
container to the script level. This allows easier mocking of the
contents and testing if the file is removed at the end.