Test for pass messages for multiple clusters
Previously, we have mocked Find to return only one cluster. Now the
successful validation of three clusters is validated.

To make the assertion easier to understand compared to directly
providing the total number of expected pass messages, arithmetic
expansion is used to specify it as the multiplication of the number of
clusters by the number of pass messages per cluster.