Replace short parameters with long ones
Short parameters have to be looked up if one is not familiar with them.
They are useful to reduce typing when using commands repeatedly in an
interactive shell.

In contrast, long parameters often allow understanding without reading
documentation. As they are not typed repeatedly in scripts and can be
split on multiple lines, they provide better readability.