Test output of header for each cluster
Tabs are ignored in such a definition of the multiline text block with a
here document, so it can be indented to distinguish it visually only in
the code.

To match the block with Grep, newline characters defined with an ANSI-C
quoting string are replaced with literal newline escape sequences. This
is done with a search-and-replace parameter expansion.