Add test for failing Find command
As we intend to refactor the search for cluster folders, we add a test
for failure to make sure that errors are correctly reported also after
making the changes.

As the search occurs at the beginning, ShellSpec's feature of asserting
the number of output lines is used to verify that nothing besides the
error message is output.