Test for clusters with newlines and spaces
The plan to refactor the cluster folder search stems from the fact that
iterating over the output of Find with a for loop as it is currently
done is error-prone.

This is because Find by default separates the found entries with a
newline character which is a valid character in filenames. The for loop
in turn splits entries on this and other valid filename characters like
spaces, so filenames could be wrongly split.

To make sure that such cases are handled correctly, a test with
two folders, one containing newlines, one spaces, is added that checks
that exactly these two are processed. As expected, it fails with the
current implementation.