Return with status of failing Find command
Errors returned by commands should be handled by exiting to avoid the
execution to continue in an undefined state.

One way to achieve this would be to set Bash's `-e` option which has
this exact function. However, it is disputed because it doesn't do so
reliably and one has to be aware of the possible catches. In addition,
it exits the script without a useful message that would allow to locate
the source of error.

Therefore, short-circuit evaluation is used as an idiomatic way in Bash
to exit upon error with a message.

For this, a function is used that first stores the last exit status as
it would be overwritten even by a variable declaration statement. The
exit message is output with the help of Printf which allows to apply one
format to several strings.